Fixes for building against macOS 10.13 SDK

+++ plugins/core/Firmata/Firmata/Connections/FirmataBluetoothLEConnection.cpp
@@ -78,7 +78,7 @@
 #if TARGET_OS_IPHONE
     if (central.state == CBManagerStatePoweredOn)
 #else
-    if (central.state == CBCentralManagerStatePoweredOn)
+    if (central.state == static_cast<NSInteger>(CBCentralManagerStatePoweredOn))
 #endif
     {
         notifyCallback(connection);
